backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pymongo==4.6.0
motor==3.3.2
pydantic==2.5.0
python-dotenv==1.0.0
google-ai-generativelanguage==0.4.0
google-generativeai==0.3.2
httpx==0.25.2
aiofiles==23.2.1
python-multipart==0.0.6

# Development dependencies
flake8==6.1.0
black==23.11.0
isort==5.12.0
numpy==1.25.2
pillow==10.1.0
requests==2.31.0
beautifulsoup4==4.12.2
tiktoken==0.5.2
typing-extensions==4.8.0